Create parent_child_translate_match.py
In my professional experience of being an animator, The one repetitive task I have encountered everywhere was to move the rigs/assets to the desired location before I started animating.

This script is my attempt to make this process faster and easier. All the user needs to do is select the set(parent object) and run the script. All the assets(children) are translated to match the co-ordinates of the set(parent).